dynamic array

+++ src/util.c
@@ -18,7 +18,7 @@ int is_palindrome( const char* str)
   size_t sz = strlen(str);
   size_t half = sz/2;
   size_t l =0u, r = sz - 1u;
-  for(int i=0; i<= half; i++)
+  for(size_t i=0; i<= half; i++)
   {
     while(!isalnum(str[l]))l++;
     while(!isalnum(str[r]))r--;
